requirements.txt
dns-python
requests
colorama